+++ example/android/app/src/main/kotlin/com/example/thwami/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.thwami
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}