+++ lib/thawani_payment.dart
@@ -1,7 +1,2 @@
 library thawani_payment;
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'widgets/pay_button.dart';