+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Adding Api
+* Adding Publishable Key As pKey
+* Adding Button To Payment
+* Adding OnCreate  
+* Adding OnPaid
+* Adding onCancelled
+* Adding onError
+* Checked From All Errors
+* Adding Button Style
+* Adding Test Mode To try the api as Test
+